+++ jupytab-server/docs/resources/ConfigSection.png
