+++ jupytab-server/docs/resources/RealEstateCrime.png
