+++ jupytab-server/docs/resources/TableauStart.png
