+++ jupytab-server/docs/resources/component-overview.png
