+++ jupytab-server/jupytab_server/log_pipe.py
