+++ jupytab-server/jupytab_server/static/favicon.ico
