+++ jupytab-server/jupytab_server/static/notebook.js
