+++ jupytab-server/requirements-dev.txt
@@ -1,3 +1,4 @@
+ipykernel
 pytest
 flake8
 pytest-cov
