+++ jupytab-server/requirements.txt
@@ -1,0 +1,1 @@
+jupyter_kernel_gateway>=2,<3
