+++ jupytab-server/samples/air-flights/AirFlights.twbx
