+++ jupytab-server/samples/config.ini
