+++ jupytab-server/samples/real-estate_crime/RealEstateCrime.twbx
