+++ jupytab-server/setup.cfg
