--- jupytab-utils/requirements.txt
@@ -1,1 +1,0 @@
-pandas<1
