--- jupytab/MANIFEST.in
@@ -1,1 +1,0 @@
-recursive-include jupytab/static *
