+++ jupytab/jupytab/__init__.py
@@ -1,0 +1,7 @@
+# Copyright (c) 2019 Capital Fund Management
+# SPDX-License-Identifier: MIT
+
+from .table import Tables
+from .dataframe_table import DataFrameTable
+
+__all__ = [Tables, DataFrameTable]
