+++ jupytab/jupytab/table.py
