+++ jupytab/requirements-dev.txt
@@ -1,4 +1,3 @@
-ipykernel
 pytest
 flake8
 pytest-cov
