+++ jupytab/setup.py
@@ -15,17 +15,8 @@ sources in a very flexible way using all the power of Python.
 This is achieved by having Tableau access data through a **web
 server created by Jupytab**.
 
-Jupytab is built on **solid foundations**: Tableau's Web Data
-Connector and the Jupyter Kernel Gateway.
-
-Features:
-
-* **Expose multiple pandas dataframes** to Tableau from a Jupyter
-  notebook
-* Access **several notebooks** from Tableau through a **single
-  entry point** (web server)
-* Manage your notebooks using a **web interface**
-* **Secure access** to your data
+This package is an helper package to split dependencies between
+Jupytab server and your notebook. It only requires pandas.
 
 The full documentation is available on the project's home page.
 """
@@ -34,7 +25,7 @@ The full documentation is available on the project's home page.
 setup(
     name="jupytab",
     version=VERSION,
-    description="Connect Tableau to your Jupyter Notebook",
+    description="Jupytab package to be used in notebooks",
     long_description_content_type="text/markdown",
     long_description=README,
     url="https://github.com/CFMTech/Jupytab",
@@ -49,15 +40,10 @@ setup(
     ],
     packages=find_packages(exclude=["*.tests", "*.tests.*"]),
     include_package_data=True,
-    install_requires=["jupyter_kernel_gateway"],
+    install_requires=["pandas"],
     project_urls={
         "Bug Tracker": "https://github.com/CFMTech/Jupytab/issues",
         "Documentation": "https://github.com/CFMTech/Jupytab",
         "Source Code": "https://github.com/CFMTech/Jupytab"
-    },
-    entry_points={
-        "console_scripts": [
-            "jupytab=jupytab.jupytab:main",
-        ]
-    },
+    }
 )
